Changed datatype from long to int in InputStreamReader

+++ FileHandlingEx/InputStreamReaderEx.java
@@ -5,13 +5,13 @@ import java.io.InputStreamReader;
 public class InputStreamReaderEx {
     public static void main(String[] args) throws IOException{
         String mobile;
-        long mobPrice;
+        int mobPrice;
         InputStreamReader isr = new InputStreamReader(System.in);
         BufferedReader br = new BufferedReader(isr);
         System.out.println("Enter your mobile company name: ");
         mobile = br.readLine();
         System.out.println("Enter your mobile number: ");
-        mobPrice = Long.parseLong(br.readLine());
+        mobPrice = Integer.parseInt(br.readLine());
         System.out.println("Your mobile company name is: "+mobile +", Your mobile price is: "+mobPrice);
     }
 }
